updated model so that the conversation id is a long, also updated the response for posting and creating a new conversation

+++ Hollerback-Android/src/com/moziy/hollerback/fragment/ConversationListFragment.java
@@ -147,7 +147,7 @@ public class ConversationListFragment extends BaseFragment {
 				long id) {
 			LogUtil.i("Starting Conversation: " + position + " id: " + id);
 
-			startConversationFragment(Integer.toString(mConversationListAdapter
+			startConversationFragment(Long.toString(mConversationListAdapter
 					.getItem((int) id).getConversation_Id()));
 
 		}
